refactor filter condition strategies and submodule comparisons
