+++ lib/rasti/db/nql/filter_condition_strategies/comparisons/sqlite/greather_than_or_equal.rb
@@ -3,17 +3,17 @@ module Rasti
     module NQL
       module FilterConditionStrategies
         module SQLiteComparisons
-          class Like < Base
+          class GreatherThanOrEqual < Comparisons::Base
             class << self
 
               def for_array(attribute, arguments)
-                Sequel.|(*arguments.map { | argument | Sequel.like(attribute, "%#{argument}%") } )
+                raise Comparisons::TypedComparisonNotSupported.new '>=', 'array'
               end
 
               private
 
               def common_filter_method(attribute, argument)
-                Sequel.ilike attribute, argument
+                attribute >= argument
               end
 
             end
