+++ lib/rasti/db/nql/filter_condition_strategies/comparisons/sqlite/like.rb
@@ -3,17 +3,17 @@ module Rasti
     module NQL
       module FilterConditionStrategies
         module SQLiteComparisons
-          class NotEqual < Base
+          class Like < Comparisons::Base
             class << self
 
               def for_array(attribute, arguments)
-                Sequel.|(*arguments.map { | argument | ~Sequel.like(attribute, "%\"#{argument}\"%") } )
+                Sequel.|(*arguments.map { | argument | Sequel.like(attribute, "%#{argument}%") } )
               end
 
               private
 
               def common_filter_method(attribute, argument)
-                Sequel.negate attribute => argument
+                Sequel.ilike attribute, argument
               end
 
             end
