+++ lib/rasti/db/nql/filter_condition_strategies/comparisons/sqlite/not_equal.rb
@@ -3,17 +3,17 @@ module Rasti
     module NQL
       module FilterConditionStrategies
         module SQLiteComparisons
-          class Include < Base
+          class NotEqual < Comparisons::Base
             class << self
 
               def for_array(attribute, arguments)
-                Sequel.|(*arguments.map { | argument | Sequel.like(attribute, "%\"#{argument}\"%") } )
+                Sequel.|(*arguments.map { | argument | ~Sequel.like(attribute, "%\"#{argument}\"%") } )
               end
 
               private
 
               def common_filter_method(attribute, argument)
-                Sequel.ilike attribute, "%#{argument}%"
+                Sequel.negate attribute => argument
               end
 
             end
